+++ CHANGELOG.md
@@ -3,11 +3,11 @@
 * Added new general settings for the Havok layer and material when importing OBJ as collision. The material can be specified as the CRC32 hash (decimal or hexadecimal with a 0x prefix), or as the string name (e.g. Broken Stone), empty defaults to None. Note that some existing settings previously under NIF have been reset because they were moved to a new Import/Export tab.
 * Resource file choosers can now default to the path of the last file of the same type (texture, material or mesh) selected, instead of the original path in the NIF or material, if the last selected file exists. This can be enabled in the general settings.
 * Fixed bugs in stitching and unstitching triangle strips, and the output is better optimized by re-stripifying the data.
+* Stripify and Triangulate convert the original geometry and data blocks instead of creating new blocks. This change allows for preserving the block order and all data fields other than the strips/triangles being converted.
 * Fixed highlighting the selected vertices of BSDynamicTriShape blocks.
 * Fixed the sort order of bhkBallSocketConstraintChain in the Reorder Blocks spell.
 * Fixed BSBlastNode, BSDamageStage and BSDebrisNode blocks being incorrectly deleted by Remove Bogus Nodes.
 * Fixed rendering the scale of bhkCompressedMeshShape, the Scale Copy field of the block (X only) now overrides the scale from parent nodes.
-* Fixed Consistency Flags not being copied by Stripify and Triangulate. Reordering blocks is still needed after using these mesh spells.
 * The texture clamp mode in BSEffectShaderProperty is shown as an enumeration instead of a byte value.
 
 #### NifSkope-2.0.dev11-20250630